Amethyst vdev DataBase.Schemas.Account

The main abstraction to a bank account per say.

Provides a set of functions to interact with it, since its creation to queries about its data. Express information over the accounts database table.

Types

final_balance_t()

final_balance_t() :: %{final_balance: Decimal.t()} | nil

A map containing a :final_balance key with a Decimal.t/0 value. Or nil.

It’s the representation for querying this specific field in the database.

response_t()

response_t() :: {:ok, t()} | {:error, any()}

A standard Ecto response to DataBase.Schemas.Account data insertion.

t()

t() :: %DataBase.Schemas.Account{
  __meta__: term(),
  api_token: term(),
  id: term(),
  incoming_transfers: term(),
  inserted_at: term(),
  movements: term(),
  name: term(),
  outgoing_transfers: term(),
  updated_at: term()
}

A DataBase.Schemas.Account struct.

Functions

activity_range(account)

activity_range(t()) :: nil | Date.Range.t()

It’s a Date.Range.t/0 representing the activity time span of a given t/0.

balance(account)

balance(t() | final_balance_t()) :: Decimal.t()

Returns the given t/0 balance.

create(account)

create(t()) :: response_t()

Tries to insert a t/0 struct into database.

credit(account, amount)

credit(t(), Decimal.t()) :: DataBase.Schemas.AccountMovement.response_t()

Registers a inbound DataBase.Schemas.AccountMovement.t/0 for a given t/0 and Decimal.t/0 as amount.

debit(account, amount)

debit(t(), Decimal.t()) :: DataBase.Schemas.AccountMovement.response_t()

Registers a outbound DataBase.Schemas.AccountMovement.t/0 for a given t/0 and Decimal.t/0 as amount.

early_balance(id, date)

early_balance(pos_integer(), Date.t()) ::
  DataBase.Schemas.AccountMovement.initial_balance_t()

Given the :id of t/0 and a Date.t/0, it determines the :initial_balance of the first DataBase.Schemas.AccountMovement.t/0 on this date.

has_funds?(account, amount)

has_funds?(t(), Decimal.t()) :: boolean()

Checks if a given t/0 has a given Decimal.t/0 as amount available in its balance.

inbounds_on(id, date)

inbounds_on(pos_integer(), Date.t()) :: Decimal.t()

A sum of all inbound DataBase.Schemas.AccountMovement.t/0 :amount for a given t/0 on a specific date.

outbounds_on(id, date)

outbounds_on(pos_integer(), Date.t()) :: Decimal.t()

A sum of all outbound DataBase.Schemas.AccountMovement.t/0 :amount for a given t/0 on a specific date.

set_token(account)

set_token(t()) :: t()

Given t/0, generates a API token to it.