Amethyst vdev DataBase.Schemas.AccountHistory

The bank account daily logbook.

It’s in charge to keep track of all DataBase.Schemas.AccountMovement.t/0 :direction and :amount of a DataBase.Schemas.Account.t/0 for each day.

It acts as logbook. Expresses information over the account_history database table. Has its :account_id and :date acting both as a composite primary key.

Provides functions to both register and report informations.

Types

query_result_t()

query_result_t() :: {:ok, Postgrex.Result.t()} | {:error, any()}

A standard Postgrex response to raw SQL queries.

Represents the registering act response. See register/1.

registering_data_t()

registering_data_t() :: [pos_integer() | Date.t() | Decimal.t()]

Represents the overall shape for the registering input.

report_range_t()

report_range_t() :: :total | Date.t() | Date.Range.t()

Represents the disposable report filters.

report_result_t()

report_result_t() :: {:ok, t()} | {:error, any()}

A standard Ecto response to DataBase.Schemas.AccountHistory data insertion.

t()

t() :: %DataBase.Schemas.AccountHistory{
  __meta__: term(),
  account: term(),
  account_id: term(),
  date: term(),
  final_balance: term(),
  inbounds: term(),
  initial_balance: term(),
  outbounds: term()
}

A DataBase.Schemas.AccountHistory struct.

Functions

register(movement)

register(DataBase.Schemas.AccountMovement.t()) ::
  DataBase.Schemas.AccountMovement.t() | :error

Registers a given DataBase.Schemas.AccountMovement.t/0 to its respective t/0 logbook properly updating it.

Creates a new one for the first DataBase.Schemas.AccountMovement.t/0 of each day.

report(account, date)

report(DataBase.Schemas.Account.t(), report_range_t()) :: report_result_t()

Given a DataBase.Schemas.Account.t/0 and a report_range_t/0, proceeds into filling a report over the requested period of time.

Returns a report_result_t/0. The report data is presented in a form of a single t/0.