Amethyst vdev DataBase.Schemas.AccountMovement

The bank account funds movimentation.

Every time a DataBase.Schemas.Account.t/0 adds or removes funds, it’s in other words: Inserting a new t/0. Which are simply called Movements. They can add (inbound) or remove (outbound) funds.

A Movement inbound/outbound is determined by its :direction (which is 1 for inbounds or -1 for outbounds). So its real amount is represented by direction * amount, as its :amount is always positive regardless of direction.

The act of registering a new movement to a DataBase.Schemas.Account.t/0 is called to move a movement.

This moving operation ensures to set required internal values and to write/update its respective DataBase.Schemas.AccountHistory.t/0 logbook.

See move/1.

Expresses information over the account_movements database table.

Types

initial_balance_t()

initial_balance_t() :: %{initial_balance: Decimal.t()} | nil

A map containing a :initial_balance key with a Decimal.t/0 value. Or nil.

It’s the representation for querying this specific field in the database.

missing_values_t()

missing_values_t() :: %{
  move_on: Date.t(),
  move_at: DateTime.t(),
  initial_balance: Decimal.t(),
  final_balance: Decimal.t()
}

A group of fields to be set after building a t/0. Prior its insertion.

precedent_assoc_t()

precedent_assoc_t() ::
  nil | %{previous_movement_id: pos_integer(), previous_movement: t()}

A group of fields that represents the association of a t/0 to the DataBase.Schemas.Account.t/0 last DataBase.Schemas.AccountMovement.t/0. A.k.a. previous movement.

response_t()

response_t() :: {:ok, t()} | {:error, any()}

A standard Ecto response to DataBase.Schemas.AccountMovement data insertion.

t()

t() :: %DataBase.Schemas.AccountMovement{
  __meta__: term(),
  account: term(),
  account_id: term(),
  amount: term(),
  direction: term(),
  final_balance: term(),
  id: term(),
  initial_balance: term(),
  inserted_at: term(),
  move_at: term(),
  move_on: term(),
  previous_movement: term(),
  previous_movement_id: term(),
  updated_at: term()
}

A DataBase.Schemas.AccountMovement struct.

Functions

build(account_id, amount, direction)

build(pos_integer(), Decimal.t(), integer()) :: t()

Builds a t/0 over the given arguments.

inbounds_on(m)

inbounds_on(t()) :: Decimal.t()

It’s a Decimal.t/0 to represent the addition of DataBase.Schemas.AccountHistory.t/0 (logbook) :inbounds_on with the given t/0 movement inbound value.

The movement inbound value is the t/0 :amount if it’s a inbound movement. Otherwise it’s zero.

initial_balance_for(movement)

initial_balance_for(t()) :: Decimal.t()

When a given t/0 is the first of its day, it returns the t/0 :initial_balance. Otherwise it returns the DataBase.Schemas.Account.early_balance/2 :initial_balance for the given t/0.

move(movement)

move(t()) :: response_t()

The only allowed way to register a t/0.

It will set its previous movement (if any), then compute its required internal values and register with the respective DataBase.Schemas.AccountHistory.t/0 logbook.

Only after registering to the logbook the t/0 is inserted.

outbounds_on(m)

outbounds_on(t()) :: Decimal.t()

It’s a Decimal.t/0 to represent the addition of DataBase.Schemas.AccountHistory.t/0 (logbook) :outbounds_on with the given t/0 movement outbound value.

The movement outbound value is the t/0 :amount if it’s a outbound movement. Otherwise it’s zero.

valid_amount?(amount)

valid_amount?(Decimal.t()) :: boolean()

Determines if a given Decimal.t/0 is greater than zero.