Amethyst vdev DataBase.Schemas.AccountTransfer

The centralized agreement for exchanging assets between two distinct accounts.

A transfer between accounts is performed by creating a outbound movement to the sender account and a inbound movement to the recipient account.

This module provides a function to register those two DataBase.Schemas.AccountMovement.t/0 as being part of a formal transfer.

See register/2.

Expresses information over the account_transfers database table.

Types

response_t()

response_t() :: {:ok, t()} | {:error, any()}

A standard Ecto response to DataBase.Schemas.AccountTransfer data insertion.

t()

t() :: %DataBase.Schemas.AccountTransfer{
  __meta__: term(),
  amount: term(),
  id: term(),
  inserted_at: term(),
  recipient_account: term(),
  recipient_account_id: term(),
  recipient_movement: term(),
  recipient_movement_id: term(),
  sender_account: term(),
  sender_account_id: term(),
  sender_movement: term(),
  sender_movement_id: term(),
  transfer_at: term(),
  updated_at: term()
}

A DataBase.Schemas.AccountTransfer struct.

Functions

register(outgoing, incoming)

register(
  DataBase.Schemas.AccountMovement.t(),
  DataBase.Schemas.AccountMovement.t()
) :: response_t()

Builds and registers a t/0.

Given the inbound and outbound DataBase.Schemas.AccountMovement.t/0, for the sender and recipient DataBase.Schemas.Account.t/0, it builds a respective t/0 and saves it.