Amethyst vdev DataBase.Services.Account.Authorize

Acts as a service to authorize a bank account API token.

Functions

perform(account_id, api_token)

perform(binary(), binary()) :: DataBase.Schemas.Account.response_t()

Returns a DataBase.Schemas.Account.response_t/0 to represent the access authorization for a given :account_id and :api_token set.