Amethyst vdev DataBase.Services.Account.Open

Acts as a service to open a new bank account.

Functions

perform(base_build)

perform(DataBase.Schemas.Account.t()) :: DataBase.Schemas.Account.response_t()

Given a DataBase.Schemas.Account.t/0, registers it in the database.

It also set a initial inbound movement of 1000.00 as a opening credit and generates its API authentication token.

It’s the standard way to open a new account.