Amethyst vdev DataBase.Services.Account.Transfer

Acts as a service to transfer assets between accounts.

Functions

perform(sender, recipient, amount)

perform(DataBase.Schemas.Account.t(), DataBase.Schemas.Account.t(), Decimal.t()) ::
  DataBase.Schemas.AccountTransfer.response_t()

Given two distinct DataBase.Schemas.Account.t/0 and a Decimal.t/0 amount, it performs the transfer.

Steps Performed

Checks if the amount is greater than zero.

Checks if the sender account has the required amount to transfer.

Debit the sender account.

Credit the recipient account.

Registers a new DataBase.Schemas.AccountTransfer.t/0.

It’s the standard way to transfer assets between accounts.