Amethyst vdev Teller.Authorize

Act as a Plug to perform authorization and load the context.

This should be our first connection layer.

Functions

call(conn, opts)

call(Plug.Conn.t(), Plug.opts()) :: Plug.Conn.t()

When the connection comes, this Plug shall seek for the authorization headers and set the authorized DataBase.Schemas.Account.t/0 into our Absinthe context.

The connection flow isn’t affected. It’s up to the Absinthe resolver to decide what to do next.

The following headers are inspected:

account-id

api-token

init(opts)

init(Plug.opts()) :: Plug.opts()

Plug initializer callback.